mqtt/requirements.txt
paho-mqtt
pandas
seaborn
kafka-python
confluent-kafka
